Apply an API interface refactoring
