+++ api/Trace/SpanContext.php
@@ -6,6 +6,8 @@ namespace OpenTelemetry\Trace;
 
 interface SpanContext
 {
+    const TRACE_FLAG_SAMPLED = 1;
+
     public function getTraceId(): string;
     public function getSpanId(): string;
     public function getTraceFlags(): int;
